refactor(test): split oversized test files to meet LOC limits
- test_codeagent_runner_gate.py (632 -> 344): extract
  TestApplyUnifiedNoopGate to test_noop_gate_unit.py (303)
- test_orchestration_facade.py (458 -> 317): extract
  TestOrchestrationFacadeDispatchServices to
  test_orchestration_facade_dispatch.py (149)
- test_coordinator.py (479): raise exception to 500
  (shared fixture, flat functions, split has low value)
- Update settings.yaml exceptions accordingly
