+++ config/settings.yaml
@@ -171,20 +171,17 @@ code_limits:
         limit: 500
         reason: "Codeagent backend 核心测试，单一测试类，刚超限，拆分收益低"
       - path: "tests/vibe3/domain/test_orchestration_facade.py"
-        limit: 450
-        reason: "单一测试类，刚超限，拆分收益低"
+        limit: 350
+        reason: "Facade 核心测试（事件发布、governance、supervisor scan），已拆出 dispatch 测试"
       - path: "tests/vibe3/execution/test_execution_lifecycle.py"
         limit: 550
         reason: "Execution lifecycle 测试集，5 个测试类围绕同一服务契约，共享 fixture"
       - path: "tests/vibe3/agents/test_codeagent_config.py"
         limit: 450
         reason: "Codeagent config 测试集，覆盖 preset 解析、fallback、migration 场景，逻辑清晰且符合测试最佳实践"
-      - path: "tests/vibe3/execution/test_codeagent_runner_gate.py"
-        limit: 500
-        reason: "Gate 核心测试，两个测试类围绕统一执行壳 gate/callback 逻辑，共享 fixture，拆分收益低"
       - path: "tests/vibe3/execution/test_coordinator.py"
-        limit: 450
-        reason: "Coordinator 核心测试，25 个测试函数围绕调度逻辑，共享 mock_dependencies fixture，拆分收益低"
+        limit: 500
+        reason: "Coordinator 核心测试，10 个测试函数围绕调度逻辑，共享 mock_dependencies fixture，拆分收益低"
 
   total_file_loc:
     v2_shell: 3000         # Shell 核心代码总行数
